pa/requirements.txt
asgiref~=3.4.1
beautifulsoup4==4.8.2
celery==5.2.2
Django==3.2.16
django-allauth==0.44.0
django-cors-headers==3.5.0
django-filter==2.4.0
django-query-builder==1.2.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
drf-yasg==1.20.0
factory_boy==2.12.0
gunicorn==20.1.0
html5lib==1.1
jsonschema==3.2.0
lxml==4.9.1
numpy==1.22
pandas~=1.2.2
psycopg2-binary==2.8.4
pyjwt==2.4.0
python-dateutil~=2.8.1
redis==3.5.3
requests~=2.25.1
scikit-learn==0.23.2
selenium==3.141.0
selenium-wire==4.4.0
